refactor flows for dashboard
