+++ frontend/admin/src/components/Layout/UserAvatar.tsx
@@ -14,7 +14,7 @@ import {
 import { 
   Logout as LogoutIcon
 } from '@mui/icons-material';
-import { useAuth } from '../../contexts/AuthContext';
+import { useAuth } from '../../hooks/useAuth';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'sonner';
