+++ frontend/admin/src/hooks/useAuth.ts
@@ -1,0 +1,20 @@
+import { createContext, useContext } from 'react';
+import type { UserData } from '../types';
+
+export interface AuthContextType {
+  user: UserData | null;
+  isAuthenticated: boolean;
+  isLoading: boolean;
+  login: (email: string, password: string) => Promise<void>;
+  logout: () => Promise<void>;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (context === undefined) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
+};
